moved plugin in remote action
