+++ bin/kuzzle
@@ -56,6 +56,7 @@ program
   .option('--remove', '*Removes the supplied plugin [name] from Kuzzle')
   .option('--activate', '*Marks the plugin as "activated" (Kuzzle ignores deactivated plugins)')
   .option('--deactivate', '*Marks the plugin as "deactivated" (Kuzzle ignores deactivated plugins)')
+  .option('--importConfig <file>', '*Import a configuration from a file for a given plugin')
   .option('-v, --npmVersion <version>', 'Installs plugin <version> from NPM (work only with --install)')
   .option('-u, --gitUrl <url>', 'Installs plugin from a GIT repo <url> (work only with --install)')
   .option('-p, --path <path>', 'Installs a plugin from directory <path> (work only with --install)')
