+++ docker-compose.yml
@@ -6,6 +6,8 @@ loadbalancer:
 
 kuzzle:
   image: kuzzleio/kuzzle
+  volumes:
+      - ".:/var/app"
   links:
     - elasticsearch
     - redis
