+++ lib/api/controllers/remoteActions/prepareDb.js
@@ -244,7 +244,10 @@ function createInternalIndex(kuzzle) {
 
   return kuzzle.pluginsManager.trigger('prepareDb:createInternalIndex', requestObject)
     .then(newRequestObject => kuzzle.workerListener.add(newRequestObject))
-    .then(() => kuzzle.indexCache.add(kuzzle.config.internalIndex));
+    .then(() => kuzzle.indexCache.add(kuzzle.config.internalIndex))
+    .catch(error => {
+      console.log('***********************', error);
+    });
 }
 
 function createRoleCollection(kuzzle, defaultRoleDefinition) {
