+++ lib/api/controllers/remoteActionsController.js
@@ -11,6 +11,7 @@ module.exports = function RemoteActionsController (kuzzle) {
     prepareDb:require('./remoteActions/prepareDb'),
     cleanAndPrepare: require('./remoteActions/cleanAndPrepare'),
     enableServices: require('./remoteActions/enableServices')
+    //pluginsManager: require('./remoteActions/managePlugins')
   };
 
   this.init = () => {
