+++ lib/api/remoteActions/index.js
@@ -17,7 +17,7 @@ module.exports = function RemoteActions (kuzzle) {
       prepareData,
       pid,
       data,
-      isPidMandatory = false;
+      isPidMandatory;
 
     if (this.actions[action] === undefined) {
       try {
@@ -53,10 +53,10 @@ module.exports = function RemoteActions (kuzzle) {
         return process.exit(1);
       }
     }
-
     this.kuzzle.config = config(params);
     // the only service we need is the internal broker
     this.kuzzle.pluginsManager = new PluginsManager(this.kuzzle);
+
     this.kuzzle.pluginsManager.init(this.kuzzle.isServer, true)
       .then(() => {
         this.kuzzle.pluginsManager.run();
@@ -68,6 +68,7 @@ module.exports = function RemoteActions (kuzzle) {
         return this.kuzzle.services.list.broker.init();
       })
       .then(() => {
+
         data = prepareData(params, args);
 
         request = new ResquestObject({controller: 'actions', action: action, body: data});
